requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
graphviz==0.20.1
importlib_resources==6.1.2
kiwisolver==1.4.5
matplotlib==3.8.3
neat-python==0.92
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pygame==2.5.2
pyparsing==3.1.1
python-dateutil==2.9.0.post0
six==1.16.0
visualize==0.5.1
zipp==3.17.0
